feat [client]: create reusable filled button
